fix: separate llm from session for circular import issue
